requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.3.6
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.0.4
django-celery-beat==2.8.0
django-environ==0.12.0
django-timezone-field==7.1
djangorestframework==3.16.0
kombu==5.5.3
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg==3.2.7
psycopg-binary==3.2.7
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.3.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
django-filter==25.1
gunicorn>=20.0.0

pytest
pytest-django
requests
